refactor: move viewModel classes into the package 'viewmodel'
